refactor(tests): reorganize tests and fix documentation
- Move 19 render tests from app/mod.rs to app/renderer.rs for better colocation
- Include orphaned rich_text_tests module to recover 17 untested cases
- Update lib.rs documentation to reflect current Component/node! macro architecture
- Replace outdated Elements factory pattern references with accurate examples
- Fix doctest compilation errors (add Default derive, mark terminal examples as no_run)
- Clean up app/mod.rs to only contain module declarations and exports

All 117 unit tests and 8 doctests now passing.
